# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SimpleBlog API Documentation",
        "description": "Implementation of Swagger with in Laravel",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/api": {},
        "/api/category": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get list of categories",
                "operationId": "327644f4d836240fa297b2822ab6ca92",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Create new category",
                "operationId": "de40ad04180a48262a05cf249f785ab3",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Technology"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Create new category succeed"
                    }
                }
            }
        },
        "/api/category/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get category by id",
                "operationId": "25690688339275caf7ded3e8702e7670",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get category by id succeed"
                    },
                    "404": {
                        "description": "Category Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Updates a category",
                "operationId": "9be67b4880d28356ade2ebc0b425415e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parameter with mutliple examples",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Delete category by id",
                "operationId": "d2437e8e0e200592efe7305af1c302d7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Delete category by id succeed"
                    }
                }
            }
        },
        "/api/comment": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Get all comment",
                "operationId": "c561bf3062eb2fd096a331aaca319c76",
                "responses": {
                    "200": {
                        "description": "Get all comment success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Create new comment",
                "operationId": "8aa076dd8e387ea517f9722d89e09af3",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "content": {
                                        "type": "string"
                                    },
                                    "post_id": {
                                        "type": "number"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "content": "Comment on post number x",
                                    "post_id": 3
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Create new comment succeed"
                    }
                }
            }
        },
        "/api/comment/{id}": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Get comment by id",
                "operationId": "b14d2a3c17aa747dc601c54bdf348bf9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get comment by id succeed"
                    },
                    "404": {
                        "description": "Comment not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Comments"
                ],
                "summary": "Updates a comment",
                "operationId": "58d5268605f3286ab81e3cf9f422a833",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parameter with mutliple examples",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "content": {
                                        "type": "string"
                                    },
                                    "post_id": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Comments"
                ],
                "summary": "Delete comment by id",
                "operationId": "141b336dc29339847c28c609722041d1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Delete comment by id succeed"
                    }
                }
            }
        },
        "/api/post": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get all posts",
                "operationId": "b4b7284c8fb22330a82699a8356a5e88",
                "responses": {
                    "200": {
                        "description": "Get all post success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Create new post",
                "operationId": "2b5846c8a1e7a22fa31af93f7accb003",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    },
                                    "user_id": {
                                        "type": "number"
                                    },
                                    "category_id": {
                                        "type": "number"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "title": "post swagger",
                                    "content": "Description about post swagger on article",
                                    "user_id": 2,
                                    "category_id": 3
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Create new post succeed"
                    }
                }
            }
        },
        "/api/post/{id}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get post by id",
                "operationId": "239ae40d24d82380dd09557f9614ad9e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get post by id succeed"
                    },
                    "404": {
                        "description": "Post Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Posts"
                ],
                "summary": "Update post",
                "operationId": "810f728ecb4e0ff17868cb92df339718",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parameter with mutliple examples",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    },
                                    "user_id": {
                                        "type": "number"
                                    },
                                    "category_id": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "Delete post by id",
                "operationId": "6b10f4215c6e3ed8be5c8f54ba69a2c3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Delete post by id succeed"
                    }
                }
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get list of users",
                "operationId": "c457726701591d1183b53aa71fc13441",
                "responses": {
                    "200": {
                        "description": "Get all user success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create new user",
                "operationId": "592819a0265360b2014512d6dbfaf0e7",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "user_swagger",
                                    "email": "swagger@gmail.com",
                                    "password": "password"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Create new user succeed"
                    }
                }
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get user by id",
                "operationId": "36a33ff774d5cba33c039dec2c3e0287",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get user by id succeed"
                    },
                    "404": {
                        "description": "User Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update user",
                "operationId": "b9091397c8b25f12c6adb74be6ce3a5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parameter with mutliple examples",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete user by id",
                "operationId": "e169ebcc01eeee2f5e57e5bfc064bc7f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Delete user by id succeed"
                    }
                }
            }
        }
    }
}